Goal-Tracker
For tracking goals. I added ActivePage, DonePage and GoalPage on the components. I also made a new page called GoalTrackerPage.

In this function currently you can add goal (title, description and tasks). when you click the In progress button it will move to the done section, indicating the goal is finished. Pop up when goal added successfully and added fail when the goal is empty. Currently still working on it to improve.